Extract component logic
